requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2026.1.4
charset-normalizer==3.4.4
Django==3.2.25
django-cors-headers==4.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.0
idna==3.10
Pillow==9.5.0
PyJWT==2.8.0
python-decouple==3.8
pytz==2025.2
requests==2.31.0
sqlparse==0.4.4
sslcommerz-lib==1.0
typing-extensions==4.7.1
urllib3==2.0.7
